+++ CHANGELOG.md
@@ -3,12 +3,17 @@
 ## 16.08.16
 - you are now able to define custom actions that will be shown as menu 
   entry in the scripting menu and optionally in the toolbar as button
-  (for [Issue #297](https://github.com/pbek/QOwnNotes/issues/297))
-    - take a look at the example at
+  (for [Issue #296](https://github.com/pbek/QOwnNotes/issues/296) and
+  for [Issue #297](https://github.com/pbek/QOwnNotes/issues/297))
+    - take a look at the examples at
       [custom-actions.qml](https://github.com/pbek/QOwnNotes/blob/develop/doc/scripting/custom-actions.qml)
+        - there is an example for opening the current note with KWrite
+          and to create a new note with a custom text
     - there is a small documentation about the `registerCustomAction`
       function at
       [README.md](https://github.com/pbek/QOwnNotes/blob/develop/doc/scripting/README.md#register-a-custom-action)
+    - you can assign shortcuts for your custom actions in the 
+      shortcut settings
 - the warning about tags when copying or moving notes to a note
   subfolder will now be only displayed if one of the selected notes has
   a tag assigned (for [Issue #289](https://github.com/pbek/QOwnNotes/issues/289))